PO-8693 : Refactor - removed paymentterms methods from ODAS
