+++ src/main/java/uk/gov/hmcts/opal/service/iface/DefendantAccountServiceInterface.java
@@ -1,13 +1,10 @@
 package uk.gov.hmcts.opal.service.iface;
 
-import uk.gov.hmcts.opal.dto.AddPaymentCardRequestResponse;
 import uk.gov.hmcts.opal.dto.DefendantAccountHeaderSummary;
-import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
 import uk.gov.hmcts.opal.dto.UpdateDefendantAccountRequest;
 import uk.gov.hmcts.opal.dto.UpdateDefendantAccountResponse;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryFilter;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryResponse;
-import uk.gov.hmcts.opal.dto.request.AddDefendantAccountPaymentTermsRequest;
 import uk.gov.hmcts.opal.dto.response.DefendantAccountAtAGlanceResponse;
 import uk.gov.hmcts.opal.dto.search.AccountSearchDto;
 import uk.gov.hmcts.opal.dto.search.DefendantAccountSearchResultsDto;
@@ -33,10 +30,4 @@ public interface DefendantAccountServiceInterface {
                                                     String postedBy) {
         return updateDefendantAccount(defendantAccountId, businessUnitId, request, postedBy, postedBy);
     }
-
-    GetDefendantAccountPaymentTermsResponse addPaymentTerms(Long defendantAccountId,
-                                            String businessUnitId,
-                                            String businessUnitUserId,
-                                            String ifMatch,
-                                            AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest);
 }
