+++ src/main/java/uk/gov/hmcts/opal/service/proxy/DefendantAccountPaymentTermsServiceProxy.java
@@ -5,6 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.opal.dto.AddPaymentCardRequestResponse;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
+import uk.gov.hmcts.opal.dto.request.AddDefendantAccountPaymentTermsRequest;
 import uk.gov.hmcts.opal.service.iface.DefendantAccountPaymentTermsServiceInterface;
 import uk.gov.hmcts.opal.service.legacy.LegacyDefendantAccountPaymentTermsService;
 import uk.gov.hmcts.opal.service.opal.DynamicConfigService;
@@ -36,5 +37,12 @@ public class DefendantAccountPaymentTermsServiceProxy implements DefendantAccoun
         String ifMatch) {
         return getCurrentModeService().addPaymentCardRequest(defendantAccountId, businessUnitId,
             businessUnitUserId, ifMatch);
+    }
+
+    @Override
+    public GetDefendantAccountPaymentTermsResponse addPaymentTerms(Long defendantAccountId, String businessUnitId,
+        String businessUnitUserId, String ifMatch, AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest) {
+        return getCurrentModeService().addPaymentTerms(defendantAccountId, businessUnitId,
+            businessUnitUserId, ifMatch, addPaymentTermsRequest);
     }
 }
