+++ src/test/java/uk/gov/hmcts/opal/service/DefendantAccountPaymentTermsServiceTest.java
@@ -1,8 +1,12 @@
 package uk.gov.hmcts.opal.service;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -10,13 +14,18 @@ import static org.mockito.Mockito.when;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.opal.authorisation.model.FinesPermission;
 import uk.gov.hmcts.opal.common.user.authorisation.exception.PermissionNotAllowedException;
 import uk.gov.hmcts.opal.common.user.authorisation.model.UserState;
+import uk.gov.hmcts.opal.controllers.util.UserStateUtil;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
+import uk.gov.hmcts.opal.dto.PaymentTerms;
+import uk.gov.hmcts.opal.dto.PostedDetails;
+import uk.gov.hmcts.opal.dto.request.AddDefendantAccountPaymentTermsRequest;
 import uk.gov.hmcts.opal.service.proxy.DefendantAccountPaymentTermsServiceProxy;
 
 @ExtendWith(MockitoExtension.class)
@@ -79,5 +88,50 @@ class DefendantAccountPaymentTermsServiceTest {
         verify(userState).anyBusinessUnitUserHasPermission(FinesPermission.SEARCH_AND_VIEW_ACCOUNTS);
         verifyNoInteractions(defendantAccountPaymentTermsServiceProxy);
         verifyNoMoreInteractions(userStateService, userState);
+    }
+
+    @Test
+    void addPaymentTerms_overwritesPostedDetailsFromUserState() {
+        Long defendantAccountId = 77L;
+        String businessUnitId = "78";
+        String ifMatch = "\"1\"";
+
+        UserState userWithPerm = UserStateUtil.permissionUser((short) 78, FinesPermission.AMEND_PAYMENT_TERMS);
+        when(userStateService.getUserStateV1FromSecurityContext()).thenReturn(userWithPerm);
+
+        AddDefendantAccountPaymentTermsRequest request = AddDefendantAccountPaymentTermsRequest.builder()
+            .paymentTerms(PaymentTerms.builder()
+                .postedDetails(PostedDetails.builder()
+                    .postedBy("FE_USER")
+                    .postedByName("FE_NAME")
+                    .build())
+                .build())
+            .build();
+
+        GetDefendantAccountPaymentTermsResponse proxyResponse = new GetDefendantAccountPaymentTermsResponse();
+        when(defendantAccountPaymentTermsServiceProxy.addPaymentTerms(eq(defendantAccountId),
+            eq(businessUnitId),
+            eq("USER01"),
+            eq(ifMatch),
+            any(AddDefendantAccountPaymentTermsRequest.class)))
+            .thenReturn(proxyResponse);
+
+        GetDefendantAccountPaymentTermsResponse result = defendantAccountPaymentTermsService.addPaymentTerms(
+            defendantAccountId, businessUnitId, ifMatch, request);
+
+        assertSame(proxyResponse, result);
+
+        ArgumentCaptor<AddDefendantAccountPaymentTermsRequest> captor =
+            ArgumentCaptor.forClass(AddDefendantAccountPaymentTermsRequest.class);
+        verify(defendantAccountPaymentTermsServiceProxy).addPaymentTerms(eq(defendantAccountId),
+            eq(businessUnitId),
+            eq("USER01"),
+            eq(ifMatch),
+            captor.capture());
+
+        PostedDetails postedDetails = captor.getValue().getPaymentTerms().getPostedDetails();
+        assertNotNull(postedDetails);
+        assertEquals("USER01", postedDetails.getPostedBy());
+        assertEquals("Normal User", postedDetails.getPostedByName());
     }
 }
